requirements.txt
streamlit==1.28.1
torch>=1.9.0
torchvision>=0.10.0
timm>=0.9.0
Pillow>=8.3.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.0.0
scikit-learn>=1.0.0

pydicom
scikit-image
huggingface_hub
opencv-python-headless
timm>=0.9.0
matplotlib
# Web app
# For optional Jupyter debugging (you can remove if not needed)
ipython
albumentations>=1.0.0
randaugment
seaborn>=0.11.0